requirements.txt
torch
torchvision
sklearn
matplotlib
numpy
scipy
tqdm
clip
ray
timm
opencv-python
Pillow
